Added Camera and onInit function for GameObject.
Camera allows you to manage which part of room should be displayed instead of displaying the full room.
Additionaly onInit function has been added for GameObject which is being executed when a new instance has been created in the room.

The example Escaping Sheeps has been updated in order to display this new feature.